+++ projects/control-service/projects/pipelines_control_service/src/main/java/com/vmware/taurus/datajobs/DeploymentModelConverter.java
@@ -5,8 +5,12 @@
 
 package com.vmware.taurus.datajobs;
 
+import com.vmware.taurus.controlplane.model.data.DataJobContacts;
+import com.vmware.taurus.controlplane.model.data.DataJobDeploymentStatus;
 import com.vmware.taurus.controlplane.model.data.DataJobResources;
+import com.vmware.taurus.controlplane.model.data.DataJobSchedule;
 import com.vmware.taurus.service.model.ActualDataJobDeployment;
+import com.vmware.taurus.service.model.DataJob;
 import com.vmware.taurus.service.model.DataJobDeploymentResources;
 import com.vmware.taurus.service.model.DesiredDataJobDeployment;
 import com.vmware.taurus.service.model.JobDeployment;
@@ -282,4 +286,48 @@ public class DeploymentModelConverter {
             : oldDeployment.getResources().getMemoryLimitMi());
     mergedDeployment.setResources(resources);
   }
+
+  public static DataJobDeploymentStatus toJobDeploymentStatus(
+      ActualDataJobDeployment actualDataJobDeployment, DataJob job) {
+    var deploymentStatus = new DataJobDeploymentStatus();
+    deploymentStatus.setJobVersion(actualDataJobDeployment.getDeploymentVersionSha());
+    deploymentStatus.setPythonVersion(actualDataJobDeployment.getPythonVersion());
+    deploymentStatus.setId(actualDataJobDeployment.getDataJobName());
+    deploymentStatus.setEnabled(actualDataJobDeployment.getEnabled());
+    deploymentStatus.setContacts(getContactsFromJob(job));
+    deploymentStatus.setSchedule(
+        new DataJobSchedule().scheduleCron(actualDataJobDeployment.getSchedule()));
+    deploymentStatus.setResources(getResourcesFromDeployment(actualDataJobDeployment));
+    deploymentStatus.setLastDeployedDate(
+        actualDataJobDeployment.getLastDeployedDate() == null
+            ? null
+            : actualDataJobDeployment.getLastDeployedDate().toString());
+    deploymentStatus.setLastDeployedBy(actualDataJobDeployment.getLastDeployedBy());
+    return deploymentStatus;
+  }
+
+  private static DataJobContacts getContactsFromJob(DataJob job) {
+    DataJobContacts contacts = new DataJobContacts();
+    if (job.getJobConfig() != null) {
+      var config = job.getJobConfig();
+      contacts.setNotifiedOnJobDeploy(config.getNotifiedOnJobDeploy());
+      contacts.setNotifiedOnJobFailurePlatformError(config.getNotifiedOnJobFailurePlatformError());
+      contacts.setNotifiedOnJobSuccess(config.getNotifiedOnJobSuccess());
+      contacts.setNotifiedOnJobFailureUserError(config.getNotifiedOnJobFailureUserError());
+    }
+    return contacts;
+  }
+
+  private static DataJobResources getResourcesFromDeployment(ActualDataJobDeployment deployment) {
+    DataJobResources resources = new DataJobResources();
+    var deploymentResources = deployment.getResources();
+    if (deploymentResources != null) {
+      resources.setCpuRequest(deploymentResources.getCpuRequestCores());
+      resources.setCpuLimit(deploymentResources.getCpuLimitCores());
+      resources.setMemoryRequest(deploymentResources.getMemoryRequestMi());
+      resources.setMemoryLimit(deploymentResources.getMemoryLimitMi());
+    }
+    return resources;
+  }
+
 }
