+++ projects/control-service/projects/pipelines_control_service/src/main/java/com/vmware/taurus/datajobs/ToModelApiConverter.java
@@ -66,49 +66,6 @@ public class ToModelApiConverter {
     return new DataJob(modelJob.getJobName(), config);
   }
 
-  public static DataJobDeploymentStatus toJobDeploymentStatus(
-      ActualDataJobDeployment actualDataJobDeployment, DataJob job) {
-    var deploymentStatus = new DataJobDeploymentStatus();
-    deploymentStatus.setJobVersion(actualDataJobDeployment.getDeploymentVersionSha());
-    deploymentStatus.setPythonVersion(actualDataJobDeployment.getPythonVersion());
-    deploymentStatus.setId(actualDataJobDeployment.getDataJobName());
-    deploymentStatus.setEnabled(actualDataJobDeployment.getEnabled());
-    deploymentStatus.setContacts(getContactsFromJob(job));
-    deploymentStatus.setSchedule(
-        new DataJobSchedule().scheduleCron(actualDataJobDeployment.getSchedule()));
-    deploymentStatus.setResources(getResourcesFromDeployment(actualDataJobDeployment));
-    deploymentStatus.setLastDeployedDate(
-        actualDataJobDeployment.getLastDeployedDate() == null
-            ? null
-            : actualDataJobDeployment.getLastDeployedDate().toString());
-    deploymentStatus.setLastDeployedBy(actualDataJobDeployment.getLastDeployedBy());
-    return deploymentStatus;
-  }
-
-  private static DataJobContacts getContactsFromJob(DataJob job) {
-    DataJobContacts contacts = new DataJobContacts();
-    if (job.getJobConfig() != null) {
-      var config = job.getJobConfig();
-      contacts.setNotifiedOnJobDeploy(config.getNotifiedOnJobDeploy());
-      contacts.setNotifiedOnJobFailurePlatformError(config.getNotifiedOnJobFailurePlatformError());
-      contacts.setNotifiedOnJobSuccess(config.getNotifiedOnJobSuccess());
-      contacts.setNotifiedOnJobFailureUserError(config.getNotifiedOnJobFailureUserError());
-    }
-    return contacts;
-  }
-
-  private static DataJobResources getResourcesFromDeployment(ActualDataJobDeployment deployment) {
-    DataJobResources resources = new DataJobResources();
-    var deploymentResources = deployment.getResources();
-    if (deploymentResources != null) {
-      resources.setCpuRequest(deploymentResources.getCpuRequestCores());
-      resources.setCpuLimit(deploymentResources.getCpuLimitCores());
-      resources.setMemoryRequest(deploymentResources.getMemoryRequestMi());
-      resources.setMemoryLimit(deploymentResources.getMemoryLimitMi());
-    }
-    return resources;
-  }
-
   public static ExecutionType toExecutionType(DataJobExecution.TypeEnum type) {
     switch (type) {
       case MANUAL:
